Fix serializers and improve their typings
Before this fix no serializer was specified when the constructor options
set one. The if would be false and nothing would set the serializer.

also fix typo in default folderPath

also used deno fmt/lint for exactly this file to ensure some formatting

This might be considered a breaking change but as it fixes non working
code of a new feature I would see this as a bugfix rather than a
breaking major.